requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
defusedxml==0.6.0
Flask==1.1.1
gspread==3.1.0
gunicorn==20.0.4
httplib2==0.15.0
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
multidict==4.7.2
numpy==1.18.1
oauth2client==4.1.3
pikepdf==1.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-forecastio==1.4.0
pytz==2019.3
requests==2.22.0
responses==0.10.9
rsa==4.0
six==1.13.0
slackclient==2.5.0
soupsieve==1.9.5
typing-extensions==3.7.4.1
urllib3==1.25.7
Werkzeug==0.16.0
yarl==1.4.2
